Create Text Analysis - Flesch Index
Requirements Write a program that computes the Flesch index and grade level for text stored in a text file. 
 
The input to this program is the name of a text file. The outputs are the number of sentences, words, and syllables in the file, as well as the file’s Flesch index and grade- level equivalent.  
 
During analysis, I consulted experts in the problem domain to learn any information that might be relevant in solving the problem. For our problem, this information includes the definitions of sentence, word, and syllable. For the purposes of this program, these terms are defined as: 
 
Word Any sequence of non- whitespace characters. Sentence Any sequence of words ending in a period, question mark, exclamation point, colon, or semicolon. Syllable Any word of three characters or less; or any vowel (a, e, i, o, u) or pair of consecutive vowels, except for a final - es, - ed, or - e that is not - le.   Note that the definitions of word and sentence are approximations. Some words, such as “ doubles” and “syllables,” end in “ es” but will be counted as having one syllable, and an ellipse (“…”) will be counted as three sentences.  
 
Flesch’s formula to calculate the index F is the following:  F = 206.835 – 1.015  * (words / sentences) – 84.6 *  (syllables / words)  
 
The Flesch- Kincaid Grade Level Formula is used to compute the Equivalent Grade Level G = 0.39 * (words / sentences) + 11.8 * (syllables / words) – 15.59 
 
** Print the index value and the grade level to the screen as well as appending to a file.  The file should contain the name of the file analyzed along with the index and grade level scores.