Create Birthday Dictionary
Birthday Dictionary: Create a program that keeps your friend’s name and birthdays in a dictionary.  Each entry in the dictionary uses a friend’s name as the key and that friend’s birthday as the value.  The program should allow you to look up your friend’s birthdays by entering their names. The program should display a menu that allows the user to make one of the following choices: 1. Look up a birthday 2. Add a new birthday 3. Change a birthday 4. Delete a birthday 5. Quit the program Create a separate function for each option.  The program should loop until the user selects the quit option.